update omsdk 1.4.7 + update omsdk js file.
(cherry picked from commit ef4ed7a5b07d09abd2aa710c506236d29fd3abe0)
