--- PubnativeLite/PubnativeLite/OMSDK-1.3.4/OMSDK_Pubnativenet.framework/Headers
@@ -1,1 +1,0 @@
-Versions/Current/Headers
