--- PubnativeLite/PubnativeLite/OMSDK-1.3.4/OMSDK_Pubnativenet.framework/OMSDK_Pubnativenet
@@ -1,1 +1,0 @@
-Versions/Current/OMSDK_Pubnativenet
