+++ PubnativeLite/PubnativeLite/OMSDK-1.3.7/OMSDK_Pubnativenet.framework/Headers
@@ -1,0 +1,1 @@
+Versions/Current/Headers
