+++ PubnativeLite/PubnativeLite/OMSDK-1.3.7/OMSDK_Pubnativenet.framework/OMSDK_Pubnativenet
@@ -1,0 +1,1 @@
+Versions/Current/OMSDK_Pubnativenet
